rename sample
