+++ feature_layers/generate-geodatabase-replica-from-feature-service/src/main/java/com/esri/samples/generate_geodatabase_replica_from_feature_service/GenerateGeodatabaseReplicaFromFeatureServiceLauncher.java
@@ -1,0 +1,11 @@
+package com.esri.samples.generate_geodatabase_replica_from_feature_service;
+
+/**
+* Wrapper required for launching a JavaFX 11 app through Gradle or from a jar.
+*/
+public class GenerateGeodatabaseReplicaFromFeatureServiceLauncher {
+
+  public static void main(String[] args) {
+    GenerateGeodatabaseReplicaFromFeatureServiceSample.main(args); 
+  }
+}
