--- feature_layers/generate-geodatabase/src/main/java/com/esri/samples/generate_geodatabase/GenerateGeodatabaseLauncher.java
@@ -1,11 +1,0 @@
-package com.esri.samples.generate_geodatabase;
-
-/**
-* Wrapper required for launching a JavaFX 11 app through Gradle or from a jar.
-*/
-public class GenerateGeodatabaseLauncher {
-
-  public static void main(String[] args) {
-    GenerateGeodatabaseSample.main(args); 
-  }
-}
